Forces int typecast for user id parameters

+++ backend/server.py
@@ -129,7 +129,7 @@ def get_metric_names():
     if valid_request(request.path, cookie, args) == False:
         return "Invalid request\n", 400
 
-    metric_names = db['Swell']['runs'].find_one({'_id': args['run_id']}, {'info': 1})
+    metric_names = db['Swell']['runs'].find_one({'_id': (int)args['run_id']}, {'info': 1})
     metric_names = set([i['name'] for i in metric_names['info']['metrics']])
 
     return encode_response(json_response(metric_names), cookie)
@@ -141,8 +141,8 @@ def get_metric_scalars():
     if valid_request(request.path, cookie, args) == False:
         return "Invalid request\n", 400
 
-    run_metric = db['Swell']['metrics'].find_one({'run_id': args['run_id']}, {'name': args['metric_name']})
-    run_metric['name'] = db['Swell']['runs'].find_one({'_id': args['run_id']}, {'config': 1})['config']['methd_tag'] + '-{}'.format(args['run_id'])
+    run_metric = db['Swell']['metrics'].find_one({'run_id': (int)args['run_id']}, {'name': args['metric_name']})
+    run_metric['name'] = db['Swell']['runs'].find_one({'_id': (int)args['run_id']}, {'config': 1})['config']['methd_tag'] + '-{}'.format(args['run_id'])
 
     return encode_response(json_response(run_metric), cookie)
 
@@ -153,7 +153,7 @@ def get_result_names():
     if valid_request(request.path, cookie, args) == False:
         return "Invalid request\n", 400
 
-    result_names = db['Swell']['runs'].find_one({'_id': args['run_id']}, {'result': 1})['result'].keys()
+    result_names = db['Swell']['runs'].find_one({'_id': (int)args['run_id']}, {'result': 1})['result'].keys()
     result_names = list(result_names)
 
     return encode_response(json_response(result_names), cookie)
@@ -165,7 +165,7 @@ def get_result_scalars():
     if valid_request(request.path, cookie, args) == False:
         return "Invalid request\n", 400
 
-    result_scalars = db['Swell']['runs'].find_one({'_id': args['run_id']}, {'result': 1})['result'][args['result_name']]
+    result_scalars = db['Swell']['runs'].find_one({'_id': (int)args['run_id']}, {'result': 1})['result'][args['result_name']]
     result_scalars = list(result_scalars)
 
     return encode_response(json_response(result_scalars), cookie)
@@ -177,7 +177,7 @@ def get_param_names():
     if valid_request(request.path, cookie, args) == False:
         return "Invalid request\n", 400
 
-    param_names = db['Swell']['runs'].find_one({'_id': args['run_id']}, {'config': 1})
+    param_names = db['Swell']['runs'].find_one({'_id': (int)args['run_id']}, {'config': 1})
     param_names = list(param_names)
 
     return encode_response(json_response(param_names), cookie)
@@ -189,7 +189,7 @@ def get_param_scalars():
     if valid_request(request.path, cookie, args) == False:
         return "Invalid request\n", 400
 
-    param_scalars = db['Swell']['runs'].find_one({'_id': args['run_id']}, {'result': 1})['result'][args['param_name']]
+    param_scalars = db['Swell']['runs'].find_one({'_id': (int)args['run_id']}, {'result': 1})['result'][args['param_name']]
     param_scalars = list(param_scalars)
 
     return encode_response(json_response(param_scalars), cookie)
@@ -203,7 +203,7 @@ def get_artifacts():
 
     fs = gridfs.GridFS(db['Swell'])
     artifacts = []
-    for file in db['Swell']['runs'].find_one({'_id': args['run_id']}, {'artifacts'})['artifacts']:
+    for file in db['Swell']['runs'].find_one({'_id': (int)args['run_id']}, {'artifacts'})['artifacts']:
         if '.png' in file['name']:
             fig = fs.get(file['file_id']).read()
     artifacts.append(fig)
